LPS-128746 Extract MockObject and MockObjectLayoutDisplayPageObjectProvider to a external class.
